requirements.txt
yt-dlp>=2023.9.24
openai-whisper>=20231106
openai>=1.3.0
googletrans==4.0.0rc1
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
pydub>=0.25.1
numpy>=1.24.0
pyaudio>=0.2.11
websockets>=11.0.3
aiohttp>=3.8.0
asyncio-throttle>=1.0.2